+++ kibitzr/checker.py
@@ -10,6 +10,7 @@ from .fetcher import (
 from .notifier import (
     SlackSession,
     TelegramBot,
+    ZapierSession,
     post_bash,
     post_gitter,
     post_mailgun,
@@ -153,6 +154,8 @@ class Checker(object):
             return SlackSession().post
         elif key == 'telegram':
             return TelegramBot().post
+        elif key == 'zapier':
+            return ZapierSession().post
         elif key == 'smtp':
             return functools.partial(post_smtp, notifier_conf=value)
         else: