+++ kibitzr/notifier/__init__.py
@@ -4,4 +4,5 @@ from .bash import post_bash  # noqa
 from .gitter import post_gitter  # noqa
 from .slack import SlackSession  # noqa
 from .telegram import TelegramBot  # noqa
+from .zapier import ZapierSession  # noqa
 from .smtp import post_smtp  # noqa